Add currencies with exponent 3 to the base gateway
The API accepts the amount to be charged as 'the integer amount of the smallest fractional', e.g. hundredts of a dollar and for these currencies, the thousandths. We follow ISO4217's definition for this. Everybody using the same definition of currency is the only sane way to interact with other gems or systems.

We've had a similar discussion in the past with whether or not TWD is a non-fractional currency. It officialy is not and has cents, but it's not used like that in practice, so some gateways do not accept a non-rounded amount. For gateways that do not follow the ISO definition the definition can be overriden, but these quirks should be isolated. See https://github.com/activemerchant/active_merchant/commit/5bb9d2c16c18afde89da20a07c3f25c7002312fe